+++ app/src/src_nonfreenet/org/breezyweather/sources/cwa/json/CwaAlertDatasetInfo.kt
@@ -20,6 +20,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CwaAlertDatasetInfo(
-    val datasetDescription: String,
+    val datasetDescription: String?,
     val validTime: CwaAlertValidTime
 )