+++ app/src/src_nonfreenet/org/breezyweather/sources/cwa/json/CwaNormalsMonthly.kt
@@ -20,7 +20,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CwaNormalsMonthly(
-    val Month: String,
-    val Maximum: String,
-    val Minimum: String
+    val Month: String?,
+    val Maximum: String?,
+    val Minimum: String?
 )