+++ app/src/src_nonfreenet/org/breezyweather/sources/cwa/json/CwaWeatherTown.kt
@@ -21,6 +21,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CwaWeatherTown(
-    @SerialName("forecast72hr") val hourly: CwaWeatherForecast,
-    @SerialName("forecastWeekday") val daily: CwaWeatherForecast
+    @SerialName("forecast72hr") val hourly: CwaWeatherForecast?,
+    @SerialName("forecastWeekday") val daily: CwaWeatherForecast?
 )